+++ yourstory/src/pages/Library.js
@@ -1,168 +1,44 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Heart, Mail } from 'lucide-react';
-import theme, { media } from '../styles/theme';
-
+import NavBar from '../components/common/NavBar';
+import LibraryHeader from '../components/library/LibraryHeader';
+import LibraryContent from '../components/library/LibraryContent';
+import { media } from '../styles/theme';
 
 const Library = () => {
-    // 더미 데이터
-    const books = [
-        {
-            id: 1,
-            title: "김금자 어르신의 이야기",
-            subtitle: "기쁨으로 맞이하는 내일",
-            likes: 78,
-            messages: 4,
-            image: "/api/placeholder/400/320",
-        },
-        {
-            id: 2,
-            title: "최명환 어르신의 이야기",
-            subtitle: "손으로 만지는 음악의 세계",
-            likes: 63,
-            messages: 2,
-            image: "/api/placeholder/400/320",
-        }
-    ];
-
-    return (
-        <Container>
-            <div>헤더</div>
-            
-            <Main>
-                <Title>이타적 자서전 도서 목록</Title>
-                
-                <BookList>
-                    {books.map((book) => (
-                        <BookCard key={book.id}>
-                            <BookInfo>
-                                <div>
-                                    <BookTitle>{book.title}</BookTitle>
-                                    <BookSubtitle>{book.subtitle}</BookSubtitle>
-                                </div>
-                                
-                                <StatsContainer>
-                                    <Stat>
-                                        <Heart />
-                                        <span>{book.likes}</span>
-                                    </Stat>
-                                    <Stat>
-                                        <Mail />
-                                        <span>{book.messages}</span>
-                                    </Stat>
-                                </StatsContainer>
-                            </BookInfo>
-                            
-                            <ImageContainer>
-                                <img src={book.image} alt={book.title} />
-                                <ViewButton 
-                                    onClick={() => window.location.href = `/book/${book.id}`}
-                                >
-                                    도서 보기
-                                </ViewButton>
-                            </ImageContainer>
-                        </BookCard>
-                    ))}
-                </BookList>
-            </Main>
-        </Container>
-    );
+  return (
+    <div>
+      <NavBar />
+      <PageContainer>
+        <LibraryHeader />
+        <Title>이타적 자서전 도서 목록</Title>
+        <LibraryContent />
+      </PageContainer>
+    </div>
+  );
 };
 
-const Container = styled.div`
-  min-height: 100vh;
-  background-color: ${({ theme }) => theme.colors.background.default};
-`;
 
-const Main = styled.main`
-  max-width: ${({ theme }) => theme.breakpoints.desktop};
-  margin: 0 auto;
-  padding: ${({ theme }) => `${theme.spacing.padding.lg} ${theme.spacing.padding.sm}`};
+const PageContainer = styled.div`
+  padding: 81px 5%;
+  margin-left: 16.5625rem;
+  font-family: Inter;
+
+  ${media.laptop} {
+    margin-left: 16.5625rem;
+    padding: 40px 20px;
+  }
+
+  ${media.tablet} {
+    padding: 20px 10px;
+    margin-left: 16.5625rem;
+  }
 `;
 
 const Title = styled.h1`
   font-size: ${({ theme }) => theme.typography.fontSize.xl};
   font-weight: ${({ theme }) => theme.typography.fontWeight.bold};
-  margin-bottom: ${({ theme }) => theme.spacing.padding.lg};
 `;
 
-const BookList = styled.div`
-  display: grid;
-  gap: ${({ theme }) => theme.spacing.padding.md};
-`;
-
-const BookCard = styled.div`
-  display: flex;
-  align-items: stretch;
-  background-color: ${({ theme }) => theme.colors.primary.main};
-  border-radius: ${({ theme }) => theme.borderRadius.xs};
-  overflow: hidden;
-  box-shadow: ${({ theme }) => theme.shadows.md};
-`;
-
-const BookInfo = styled.div`
-  flex-grow: 1;
-  padding: ${({ theme }) => theme.spacing.padding.md};
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-`;
-
-const BookTitle = styled.h2`
-  color: ${({ theme }) => theme.colors.text.white};
-  font-size: ${({ theme }) => theme.typography.fontSize.md};
-  margin-bottom: ${({ theme }) => theme.spacing.padding.xs};
-`;
-
-const BookSubtitle = styled.h3`
-  color: ${({ theme }) => theme.colors.text.white};
-  font-size: ${({ theme }) => theme.typography.fontSize.xl};
-  font-weight: ${({ theme }) => theme.typography.fontWeight.bold};
-  margin-bottom: ${({ theme }) => theme.spacing.padding.sm};
-`;
-
-const StatsContainer = styled.div`
-  display: flex;
-  align-items: center;
-  gap: ${({ theme }) => theme.spacing.padding.md};
-`;
-
-const Stat = styled.div`
-  display: flex;
-  align-items: center;
-  color: ${({ theme }) => theme.colors.text.white};
-  
-  svg {
-    width: 1.25rem;
-    height: 1.25rem;
-    margin-right: ${({ theme }) => theme.spacing.padding.xs};
-  }
-`;
-
-const ImageContainer = styled.div`
-  width: 20rem;
-  position: relative;
-  
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-`;
-
-const ViewButton = styled.button`
-  position: absolute;
-  bottom: ${({ theme }) => theme.spacing.padding.sm};
-  right: ${({ theme }) => theme.spacing.padding.sm};
-  background-color: ${({ theme }) => theme.colors.primary.main};
-  color: ${({ theme }) => theme.colors.text.white};
-  padding: ${({ theme }) => `${theme.spacing.padding.xs} ${theme.spacing.padding.md}`};
-  border-radius: ${({ theme }) => theme.borderRadius.pill};
-  transition: ${({ theme }) => theme.transitions.short};
-  
-  &:hover {
-    background-color: ${({ theme }) => theme.colors.primary.dark};
-  }
-`;
 
 export default Library;
